modify the configured item imagepullpolicy in the dashboard template file
If the configured item cimages_repo is assigned, the docker image was supposed to be downloaded from the specified repo. But if the configured item imagepullpolicy is assigned to "always", the cimages_repo do not work well, the docker image will still be pulled from google. So I set imagepullpolicy to IfNotPresent, so that we can make the cimages_repo to take effect.

Change-Id: I613c03ae3d2409ab09e083564909b557f68f7f27