modify the setting imagepullpolicy in the dashboard j2 file
If the setting cimages_repo is assigned,
the docker image was supposed to be downloaded from the specified repo.
But if the setting imagepullpolicy is assigned to "always",
the docker image will still be pulled from google.
So I set imagepullpolicy to IfNotPresent,
so that we can make the cimages_repo to take effect.

Change-Id: I9f534e5481a5006e2ae6c3f25df55c80e1a7cf2c