Replace floating ip references
Floating IPs as a concept should not be a goal. They are an
implementation detail and one that is not universally desirable. There
are clouds that allow the user to directly attachment VMs onto public networks
rather than forcing them through NAT, which is a desirable choice and
not one that should be discouraged in any way.

"auto_ip" in the ansible playbooks will do the right thing. It will find
the right networks and use them and will get a floating IP if needed,
but will correctly not do so otherwise.

If auto_ip is used, delete_fip can be true and the ip_flag logic is not
needed.

Change-Id: Ic4e79f845267685d932eaadc1802d0e664e27b7b