Use interface_ip
interface_ip contains the best ip to use to connect to a given server.
Also, remove the logic to use private_ip if public_ip is absent. If the
environment is one where the caller is on the same private network as
the server is being created on and private=True is set, interface_ip
will contain the private address.

Leave the logic that sets private_v4 to public_v4 if private_v4 doesn't
exist, becauase it's perfectly fine for a server to not have a private
address and the public address can fill in for it just fine. But set it
to interface_ip instead in case the cloud is providing IPv6.

Change-Id: I2f41fc056215c8c8ecb6c278fad8da546ddc0d55